+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feat: CI now builds `.aab` files needed for Google Play Store release, generating two variants (`rich` and `lite`) analogous to APK builds.
 - Feat: Virtual Scrolling of messages in conversations page, to ensure smooth scrolling of a large collection of already-loaded messages.
 - Fix: Avoid rendering too many messages at once within a conversation, which could cause the app to crash.
+- Fix: Remove bottom loading spinner from message list when scrolling, and implement concurrent message loading.
 
 ## [0.7.5] - 2025-01-10